Update the dropdown menu to me more reusable and update header to have correct format
